tests: drivers: i2c: add generic sensor test with DT auto-detection
Replace per-sensor test file with a single test_i2c_sensor.c that
uses DT_HAS_COMPAT_STATUS_OKAY to compile tests only for sensors
present in the devicetree. Core controller/target sources are
excluded for sensor-only builds so the i2c_target alias is not
required.

Each sensor overlay is self-contained with its own i2c-controller
alias, avoiding conflicts with the core i2c.overlay. To add a new
sensor, add a DT_HAS_COMPAT_STATUS_OKAY block in test_i2c_sensor.c,
a DT overlay, and a testcase.yaml entry.
