+++ tests/drivers/i2c/boards/i2c_sensor.overlay
