+++ tests/drivers/i2c/testcase.yaml
@@ -111,7 +111,7 @@ tests:
       - CONFIG_I2C_HARDWARE_AWARE_TESTS=y
       - CONFIG_I2C_TARGET_BUFFER_MODE=y
 
-  drivers.i2c.bme680:
+  drivers.i2c.sensor:
     tags:
       - drivers
       - i2c
@@ -119,9 +119,8 @@ tests:
     depends_on:
       - i2c
       - sensor
-    filter: dt_alias_exists("i2c_controller")
     extra_dtc_overlay_files:
-      - boards/i2c.overlay
-      - boards/i2c_bme680.overlay
+      - boards/i2c_sensor.overlay
     extra_configs:
-      - CONFIG_I2C_BME680_TESTS=y
+      - CONFIG_I2C_SENSOR_TESTS=y
+      - CONFIG_SENSOR=y
